Revert "Cluster report: Refactor and adding tests (#292)"
This reverts commit b1f5709fb25ac7f4e6dad32c76671f1bab126f63.
