+++ pkg/transform/cluster_report_transform.go
@@ -2,8 +2,11 @@ package transform
 
 import (
 	"github.com/fusor/cpma/pkg/api"
-	"github.com/fusor/cpma/pkg/transform/cluster"
+	O7tapiroute "github.com/openshift/api/route/v1"
 	"github.com/sirupsen/logrus"
+
+	k8sapicore "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/api/resource"
 )
 
 // ClusterReportName is the cluster report name
@@ -22,7 +25,7 @@ type ClusterTransform struct {
 func (e ClusterReportExtraction) Transform() ([]Output, error) {
 	logrus.Info("ClusterTransform::Transform")
 
-	clusterReport := cluster.GenClusterReport(api.Resources{
+	clusterReport := genClusterReport(api.Resources{
 		PersistentVolumeList: e.PersistentVolumeList,
 		StorageClassList:     e.StorageClassList,
 		NamespaceList:        e.NamespaceList,
@@ -35,6 +38,160 @@ func (e ClusterReportExtraction) Transform() ([]Output, error) {
 
 	outputs := []Output{output}
 	return outputs, nil
+}
+
+// genClusterReport inserts report values into structures for json output
+func genClusterReport(apiResources api.Resources) (clusterReport ClusterReport) {
+	clusterReport.reportNodes(apiResources)
+	clusterReport.reportNamespaces(apiResources)
+	clusterReport.reportPVs(apiResources)
+	clusterReport.reportStorageClasses(apiResources)
+	return
+}
+
+// reportNodes fills in information about nodes
+func (clusterReport *ClusterReport) reportNodes(apiResources api.Resources) {
+	logrus.Debug("ClusterReport::ReportNodes")
+
+	for _, node := range apiResources.NodeList.Items {
+		nodeReport := &NodeReport{
+			Name: node.ObjectMeta.Name,
+		}
+
+		isMaster, ok := node.ObjectMeta.Labels["node-role.kubernetes.io/master"]
+		nodeReport.MasterNode = ok && isMaster == "true"
+
+		reportNodeResources(nodeReport, node.Status, apiResources)
+		clusterReport.Nodes = append(clusterReport.Nodes, *nodeReport)
+	}
+}
+
+// reportResources parse and insert info about consumed resources
+func reportNodeResources(repotedNode *NodeReport, nodeStatus k8sapicore.NodeStatus, apiResources api.Resources) {
+	repotedNode.Resources.CPU = nodeStatus.Capacity.Cpu()
+
+	repotedNode.Resources.MemoryCapacity = nodeStatus.Capacity.Memory()
+
+	memConsumed := new(resource.Quantity)
+	memCapacity, _ := nodeStatus.Capacity.Memory().AsInt64()
+	memAllocatable, _ := nodeStatus.Allocatable.Memory().AsInt64()
+	memConsumed.Set(memCapacity - memAllocatable)
+	memConsumed.Format = resource.BinarySI
+	repotedNode.Resources.MemoryConsumed = memConsumed
+
+	var runningPodsCount int64
+	for _, resources := range apiResources.NamespaceList {
+		for _, pod := range resources.PodList.Items {
+			if pod.Spec.NodeName == repotedNode.Name {
+				runningPodsCount++
+			}
+		}
+	}
+	podsRunning := new(resource.Quantity)
+	podsRunning.Set(runningPodsCount)
+	podsRunning.Format = resource.DecimalSI
+	repotedNode.Resources.RunningPods = podsRunning
+
+	repotedNode.Resources.PodCapacity = nodeStatus.Capacity.Pods()
+}
+
+// reportNamespaces fills in information about Namespaces
+func (clusterReport *ClusterReport) reportNamespaces(apiResources api.Resources) {
+	logrus.Debug("ClusterReport::ReportNamespaces")
+
+	for _, resources := range apiResources.NamespaceList {
+		reportedNamespace := NamespaceReport{Name: resources.NamespaceName}
+
+		reportPods(&reportedNamespace, resources.PodList)
+		reportResources(&reportedNamespace, resources.PodList)
+		reportRoutes(&reportedNamespace, resources.RouteList)
+		clusterReport.Namespaces = append(clusterReport.Namespaces, reportedNamespace)
+	}
+}
+
+// reportPods creates info about cluster pods
+func reportPods(reportedNamespace *NamespaceReport, podList *k8sapicore.PodList) {
+	for _, pod := range podList.Items {
+		reportedPod := &PodReport{Name: pod.Name}
+		reportedNamespace.Pods = append(reportedNamespace.Pods, *reportedPod)
+
+		// Update namespace touch timestamp
+		if pod.ObjectMeta.CreationTimestamp.Time.Unix() > reportedNamespace.LatestChange.Time.Unix() {
+			reportedNamespace.LatestChange = pod.ObjectMeta.CreationTimestamp
+		}
+	}
+}
+
+func reportResources(reportedNamespace *NamespaceReport, podList *k8sapicore.PodList) {
+	resources := ContainerResourcesReport{
+		CPUTotal:    &resource.Quantity{Format: resource.DecimalSI},
+		MemoryTotal: &resource.Quantity{Format: resource.BinarySI},
+	}
+	reportedNamespace.Resources = resources
+
+	for _, pod := range podList.Items {
+		reportContainerResources(reportedNamespace, &pod)
+	}
+}
+
+func reportRoutes(reportedNamespace *NamespaceReport, routeList *O7tapiroute.RouteList) {
+	for _, route := range routeList.Items {
+		reportedRoute := &RouteReport{
+			Name:              route.Name,
+			AlternateBackends: route.Spec.AlternateBackends,
+			Host:              route.Spec.Host,
+			Path:              route.Spec.Path,
+			To:                route.Spec.To,
+			TLS:               route.Spec.TLS,
+			WildcardPolicy:    route.Spec.WildcardPolicy,
+		}
+
+		reportedNamespace.Routes = append(reportedNamespace.Routes, *reportedRoute)
+	}
+}
+func reportContainerResources(reportedNamespace *NamespaceReport, pod *k8sapicore.Pod) {
+	cpuTotal := reportedNamespace.Resources.CPUTotal.Value()
+	memoryTotal := reportedNamespace.Resources.MemoryTotal.Value()
+
+	for _, container := range pod.Spec.Containers {
+		cpuTotal += container.Resources.Requests.Cpu().Value()
+		memoryTotal += container.Resources.Requests.Memory().Value()
+	}
+	reportedNamespace.Resources.CPUTotal.Set(cpuTotal)
+	reportedNamespace.Resources.MemoryTotal.Set(memoryTotal)
+	reportedNamespace.Resources.ContainerCount += len(pod.Spec.Containers)
+}
+
+func (clusterReport *ClusterReport) reportPVs(apiResources api.Resources) {
+	logrus.Debug("ClusterReport::ReportPVs")
+	pvList := apiResources.PersistentVolumeList
+
+	// Go through all PV and save required information to report
+	for _, pv := range pvList.Items {
+		reportedPV := &PVReport{
+			Name:         pv.Name,
+			Driver:       pv.Spec.PersistentVolumeSource,
+			StorageClass: pv.Spec.StorageClassName,
+			Capacity:     pv.Spec.Capacity,
+			Phase:        pv.Status.Phase,
+		}
+
+		clusterReport.PVs = append(clusterReport.PVs, *reportedPV)
+	}
+}
+
+func (clusterReport *ClusterReport) reportStorageClasses(apiResources api.Resources) {
+	logrus.Debug("ClusterReport::ReportStorageClasses")
+	// Go through all storage classes and save required information to report
+	storageClassList := apiResources.StorageClassList
+	for _, storageClass := range storageClassList.Items {
+		reportedStorageClass := &StorageClassReport{
+			Name:        storageClass.Name,
+			Provisioner: storageClass.Provisioner,
+		}
+
+		clusterReport.StorageClasses = append(clusterReport.StorageClasses, *reportedStorageClass)
+	}
 }
 
 // Validate no need to validate it, data is exctracted from API
