--- pkg/transform/testdata/expected-report-cluster.json
@@ -1,82 +1,0 @@
-{
- "cluster": {
-  "nodes": [
-   {
-    "name": "test-master",
-    "masterNode": true,
-    "resources": {
-     "cpu": "2",
-     "memoryConsumed": "2Ki",
-     "memoryCapacity": "2Ki",
-     "runningPods": "0",
-     "podCapacity": "10"
-    }
-   }
-  ],
-  "namespaces": [
-   {
-    "name": "testNamespace",
-    "latestChange": "2009-11-17T21:34:58Z",
-    "resources": {
-     "containerCount": 0,
-     "cpuTotal": "0",
-     "memoryTotal": "0"
-    },
-    "pods": [
-     {
-      "name": "test-pod1"
-     },
-     {
-      "name": "test-pod2"
-     }
-    ],
-    "routes": [
-     {
-      "name": "route1",
-      "host": "testhost",
-      "path": "testpath",
-      "alternateBackends": [
-       {
-        "kind": "testkind",
-        "name": "testname",
-        "weight": null
-       }
-      ],
-      "tls": {
-       "termination": "edge"
-      },
-      "to": {
-       "kind": "testkindTo",
-       "name": "testTo",
-       "weight": null
-      },
-      "wildcardPolicy": "None"
-     }
-    ]
-   }
-  ],
-  "pvs": [
-   {
-    "name": "testpv",
-    "driver": {
-     "nfs": {
-      "server": "example.com",
-      "path": ""
-     }
-    },
-    "storageClass": "testclass",
-    "capacity": {
-     "cpu": "1",
-     "memory": "1"
-    },
-    "phase": "Pending"
-   }
-  ],
-  "storageClasses": [
-   {
-    "name": "testclass",
-    "provisioner": "testprovisioner"
-   }
-  ]
- }
-}
